refactor: extract worker message handling into applyMappingsCore
